Move tab-data and tab-taiwan to tabs
